+++ src/software/ai/hl/stp/tactic/patrol_tactic.cpp
@@ -2,7 +2,6 @@
 
 #include "software/ai/hl/stp/action/move_action.h"
 #include "software/ai/hl/stp/action/stop_action.h"
-#include "software/ai/hl/stp/tactic/mutable_tactic_visitor.h"
 #include "software/logger/logger.h"
 
 
@@ -77,7 +76,8 @@ void PatrolTactic::calculateNextAction(ActionCoroutine::push_type &yield)
         if (move_action->done())
         {
             move_action->restart();
-            patrol_point_index = (patrol_point_index + 1) % patrol_points.size();
+            patrol_point_index = (patrol_point_index + 1) %
+                                 static_cast<unsigned int>(patrol_points.size());
             move_action->updateControlParams(
                 *robot, patrol_points.at(patrol_point_index),
                 orientation_at_patrol_points, linear_speed_at_patrol_points,
